+++ src/Traits/DefaultOrder/DefaultOrderBehavior.php
@@ -7,8 +7,6 @@ use Efabrica\NetteRepository\Traits\RepositoryBehavior;
 
 class DefaultOrderBehavior extends RepositoryBehavior
 {
-    private const COLUMNS_PATTERN = '/(?<!\.)\b(?!ASC\b)(?!DESC\b)([A-Za-z_][A-Za-z0-9_]*)\b(?!\s*\.)/i';
-
     private string $columns;
 
     private array $params;
@@ -25,10 +23,7 @@ class DefaultOrderBehavior extends RepositoryBehavior
     public function apply(QueryInterface $query): void
     {
         if ($query->getOrder() === []) {
-            $query->order(
-                preg_replace(self::COLUMNS_PATTERN, $query->getName() . '.$1', $this->columns),
-                ...$this->params
-            );
+            $query->order($this->columns, ...$this->params);
         }
     }
 }